#1 - Fixed refactor, disabled gifted messenger for now since it has to reside in recent navigator instead
